Move sorting functions to utils
