+++ packages/ui-components/src/components/CompareTable/index.ts
@@ -3,3 +3,4 @@ export * from "./CompareTable.style";
 export * from "./SortControls";
 export * from "./ColumnHeader";
 export * from "./interfaces";
+export * from "./utils";
